Create sum of all the even numbers
Write a C program to calculate the sum of all the even numbers between 1 and N (inclusive), where N is a positive integer provided by the user. Ensure that your program handles invalid input gracefully.

Your program should have the following specifications:

Prompt the user to enter a positive integer N.
If N is less than 1, print an error message and terminate the program.
Calculate the sum of all even numbers between 1 and N (inclusive).
Print the calculated sum.
Provide the solution to this problem in the form of a C program.