Create check if a given number is a palindrome.
Write a C program to check if a given number is a palindrome. A palindrome number is a number that remains the same when its digits are reversed. The program should prompt the user to enter a number and then print whether it is a palindrome or not.

Your program should have the following specifications:

Prompt the user to enter a number.
Read the number input from the user.
Check if the given number is a palindrome.
Print whether the number is a palindrome or not.
Provide the solution to this problem in the form of a C program.